Revert "include 'smolscale' lib for resize and swizzle"
This reverts commit 9f75fe904e84fc53d34d89de54f92272578b62df.

+++ src/nr-blit.c
@@ -84,7 +84,7 @@ void blit_core_(uint32_t *dst, int x, int y, int dst_width, int dst_height,
       memcpy(
         dst + (y + yoff) * dst_width + x, 
         src + y1 * src_width + x0, 
-        (unsigned long)w * sizeof(int32_t)
+        w * sizeof(int32_t)
       );
     }
   }
